requirements.txt
Flask
numpy
Pillow
opencv-python-headless
gunicorn
